Move saving function to Sensor
Now saving can be run headlessly on per station basis
