+++ test/test_dat_file_save.py
@@ -10,8 +10,8 @@ import numpy as np
 import scipy.io as sio
 
 import filtering as filt
-from main import load_station_data, assemble_ts_text, save_ts_files, save_mat_high_fq, save_fast_delivery
-from my_widgets import find_outliers, get_top_level_directory
+from main import load_station_data
+from my_widgets import find_outliers
 
 dirname = os.path.dirname(__file__)
 input_filename = os.path.join(dirname, 'test_data/monp/ssaba1810.dat')
@@ -44,17 +44,17 @@ class TestDatFileSave(unittest.TestCase):
         # for the two files and they are completely equal. So this is essentially an end to end test (sort of)
         # 5) Compare that formatting is exactly the same, it's all strings after all
 
-        text_data_input = assemble_ts_text(self.input_data)
-        text_data_truth = assemble_ts_text(self.data_truth)
+        text_data_input = self.input_data.assemble_ts_text()
+        text_data_truth = self.input_data.assemble_ts_text()
         self.assertEqual(text_data_input[0][1], text_data_truth[0][1])
 
         # test ts file save
         # Compare the saved file to the ground truth file
         with tempfile.TemporaryDirectory() as tmp:
-            success, failure = save_ts_files(text_data_input, tmp)
+            success, failure = self.input_data.save_ts_files(text_data_input, tmp)
             self.assertEqual(len(success), 1)
             save_folder = "t123"
-            save_path = get_top_level_directory(parent_dir=tmp) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
+            save_path = st.get_top_level_directory(parent_dir=tmp) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
                 2018)
             self.assertEqual(success[0]['message'], 'Success \nt1231810.dat Saved to ' + str(save_path) + '\n')
             self.assertEqual(success[0]['title'], 'Success')
@@ -75,15 +75,15 @@ class TestDatFileSave(unittest.TestCase):
         station_data = load_station_data([file1, file2, file3])
         # station_data_truth = load_station_data([truth_file1, truth_file2, truth_file3])
 
-        data_as_text = assemble_ts_text(station_data)
+        data_as_text = station_data.assemble_ts_text()
         # data_as_text_truth = assemble_ts_text(station_data_truth)
 
         # Compare the saved file to the ground truth file
         with tempfile.TemporaryDirectory() as tmp:
             save_folder = "t123"
-            save_path = get_top_level_directory(parent_dir=tmp) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
+            save_path = st.get_top_level_directory(parent_dir=tmp) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
                 2018)
-            success, failure = save_ts_files(data_as_text, tmp)
+            success, failure = station_data.save_ts_files(data_as_text, tmp)
             self.assertEqual(len(success), 3)
             self.assertEqual(success[0]['message'], 'Success \nt1231809.dat Saved to ' + str(save_path) + '\n')
             self.assertEqual(success[1]['message'], 'Success \nt1231810.dat Saved to ' + str(save_path) + '\n')
@@ -106,9 +106,9 @@ class TestDatFileSave(unittest.TestCase):
 
         with tempfile.TemporaryDirectory() as tmp_dir:
             save_folder = "t123"
-            save_path = get_top_level_directory(parent_dir=tmp_dir) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
+            save_path = st.get_top_level_directory(parent_dir=tmp_dir) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
                 2018)
-            save_mat_high_fq(station, tmp_dir, callback=None)
+            station.save_mat_high_fq(tmp_dir, callback=None)
             for month in station.month_collection:
                 # Compare every sensor (one file per sensor)
                 for key, sensor in month.sensor_collection.items():
@@ -151,10 +151,10 @@ class TestDatFileSave(unittest.TestCase):
 
         with tempfile.TemporaryDirectory() as tmp_dir:
             save_folder = "t123"
-            save_path = get_top_level_directory(parent_dir=tmp_dir) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
+            save_path = st.get_top_level_directory(parent_dir=tmp_dir) / st.HIGH_FREQUENCY_FOLDER / save_folder / str(
                 2018)
             clean_station.back_propagate_changes(clean_station.aggregate_months['data'])
-            save_mat_high_fq(clean_station, tmp_dir, callback=None)
+            clean_station.save_mat_high_fq(tmp_dir, callback=None)
             for month in clean_station.month_collection:
                 # Compare every sensor (one file per sensor)
                 for key, sensor in month.sensor_collection.items():
@@ -206,9 +206,9 @@ class TestDatFileSave(unittest.TestCase):
 
         with tempfile.TemporaryDirectory() as tmp_dir:
             save_folder = "t123"
-            save_path = get_top_level_directory(parent_dir=tmp_dir) / st.FAST_DELIVERY_FOLDER / save_folder / str(
+            save_path = st.get_top_level_directory(parent_dir=tmp_dir) / st.FAST_DELIVERY_FOLDER / save_folder / str(
                 2018)
-            save_fast_delivery(station, path=tmp_dir, din_path=DIN, callback=None)
+            station.save_fast_delivery(path=tmp_dir, din_path=DIN, callback=None)
             # .mat files test
             # Hourly test
             data = sio.loadmat(os.path.join(save_path, 'th1231809.mat'))
